delete TransactionMapper and apply direct call to controller

+++ transaction-service/src/main/resources/db/migration/V2__create_account_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE account
 (
-    id              BIGSERIAL PRIMARY KEY,
+    id              BIGINT PRIMARY KEY,
     balance         DECIMAL,
     account_number  VARCHAR(20),
     currency        VARCHAR(2),
-    user_id         INT REFERENCES users(id) ON DELETE SET NULL
+    user_id         BIGINT REFERENCES users(id) ON DELETE SET NULL
 );
